#2 : database design is complete for group and implementation from frontend to database is complete. Default user is added in the group admin as of now.

+++ src/main/sql/group_users.sql
@@ -1,6 +1,6 @@
 CREATE TABLE group_users (
   group_id int NOT NULL,
-  user_id BIGINT(20) NOT NULL,
+  user_id int NOT NULL,
   active varchar(30) DEFAULT NULL,
   isAdmin boolean not null default 0,
   PRIMARY KEY (group_id, user_id),
